requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.7,<3.0
networkx>=3.1,<4.0
tiktoken>=0.5.1
openai
anthropic
google-generativeai>=0.3.0
packaging>=23.2,<24.0
tenacity>=8.2.3

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
pytest-xdist>=3.3.1
pytest-benchmark>=4.0.0
pytest-html>=4.1.1
pytest-metadata>=3.0.0
pytest-randomly>=3.15.0

# Utility dependencies
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.27,<0.29
aiohttp>=3.8.0

# Development dependencies
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
structlog>=24.2,<25.0
opentelemetry-api>=1.24,<2.0
opentelemetry-sdk>=1.24,<2.0 